move GridArrows into its own file
